feat: refine app shell and PWA mobile flow
Extract the authenticated app shell from App.vue, keep auth routes outside the product shell, and improve the PWA/mobile pass.

Changes:
- Add AppShell for header, navigation, notifications and logout
- Keep VitePWA as the single service worker registration owner
- Import push-sw.js into generated Workbox service workers
- Remove stale public Workbox artifact
- Improve PWA install prompt timer cleanup and mobile safe-area behavior
- Add PWA/mobile verification notes

Verification:
- npm run type-check
- npm run build-only
- dist/sw.js imports push-sw.js and dist/push-sw.js exists
- no manual serviceWorker.register or /sw.js source refs
- no public workbox-* files remain
- git diff --check
- dash scan passed for app shell and pwa changed files
